+++ src/Messages/AccountQuery.php
@@ -2,14 +2,8 @@
 
 namespace Abivia\Ledger\Messages;
 
-use Abivia\Ledger\Exceptions\Breaker;
-use Abivia\Ledger\Helpers\Merge;
-use Abivia\Ledger\Models\LedgerAccount;
-use Abivia\Ledger\Messages\Message;
-use Carbon\Carbon;
-use TypeError;
-
-class AccountQuery extends Message {
+class AccountQuery extends Paginated
+{
 
     /**
      * @var EntityRef For pagination, a reference to the last account in the previous page.
@@ -17,7 +11,9 @@ class AccountQuery extends Message {
     public EntityRef $after;
 
     protected static array $copyable = [
-        'limit'
+        'limit',
+        'range',
+        'rangeEnding',
     ];
 
     /**
@@ -26,27 +22,11 @@ class AccountQuery extends Message {
     public EntityRef $domain;
 
     /**
-     * @var int The maximum number of accounts to return.
-     */
-    public int $limit;
-
-    /**
-     * @var string The code for the first account.
-     */
-    public string $range;
-
-    /**
-     * @var string The code for the last account.
-     */
-    public string $rangeEnding;
-
-    /**
      * @inheritDoc
      */
     public static function fromArray(array $data, int $opFlags = self::OP_ADD): self
     {
-        $query = new AccountQuery();
-        $query->copy($data, $opFlags);
+        $query = parent::fromArray($data, $opFlags);
         if (isset($data['after'])) {
             $query->after = EntityRef::fromArray($data['after'], $opFlags);
         }
@@ -57,20 +37,4 @@ class AccountQuery extends Message {
         return $query;
     }
 
-    /**
-     * @inheritDoc
-     */
-    public function validate(int $opFlags = 0): self
-    {
-        // Limit results on API calls
-        if ($opFlags & self::F_API) {
-            $limit = LedgerAccount::rules()->pageSize;
-            if (isset($this->limit)) {
-                $this->limit = min($this->limit, $limit);
-            } else {
-                $this->limit = $limit;
-            }
-        }
-        return $this;
-    }
 }